+++ packages/engine/Source/Scene/Cesium3DTileset.js
@@ -107,6 +107,7 @@ import ImageryLayerCollection from "./ImageryLayerCollection.js";
  * @property {ClippingPolygonCollection} [clippingPolygons] The {@link ClippingPolygonCollection} used to selectively disable rendering the tileset.
  * @property {ClassificationType} [classificationType] Determines whether terrain, 3D Tiles or both will be classified by this tileset. See {@link Cesium3DTileset#classificationType} for details about restrictions and limitations.
  * @property {HeightReference} [heightReference] Sets the {@link HeightReference} for features in vector tilesets.
+ * @property {number} [vectorZIndex=0] Integer z-order for vector tiles, used to "layer" primitives at the same depth and to prevent z-fighting.
  * @property {Scene} [scene] The {@link CesiumWidget#scene} that the tileset will be rendered in, required for tilesets that specify a {@link heightReference} value for clamping 3D Tiles vector data content- like points, lines, and labels- to terrain or 3D tiles.
  * @property {Ellipsoid} [ellipsoid=Ellipsoid.WGS84] The ellipsoid determining the size and shape of the globe.
  * @property {object} [pointCloudShading] Options for constructing a {@link PointCloudShading} object to control point attenuation based on geometric error and lighting.
@@ -346,6 +347,7 @@ function Cesium3DTileset(options) {
 
   this._classificationType = options.classificationType;
   this._heightReference = options.heightReference;
+  this._vectorZIndex = options.vectorZIndex ?? 0;
   this._scene = options.scene;
 
   //>>includeStart('debug', pragmas.debug);
@@ -1891,6 +1893,24 @@ Object.defineProperties(Cesium3DTileset.prototype, {
   heightReference: {
     get: function () {
       return this._heightReference;
+    },
+  },
+
+  /**
+   * Integer z-order for vector tiles, used to "layer" primitives at the same depth
+   * and to prevent z-fighting.
+   *
+   * @memberof Cesium3DTileset.prototype
+   *
+   * @type {number}
+   * @default 0
+   *
+   * @experimental This feature is not final and is subject to change without Cesium's standard deprecation policy.
+   * @readonly
+   */
+  vectorZIndex: {
+    get: function () {
+      return this._vectorZIndex;
     },
   },